refactor(dev-core): split oversized migrate.ts + doctor.ts (#196)
Behavior-preserving decomposition flagged by the dev-core maintainability audit:

- migrate.ts (898 l) → migrate-{shared,audit,backfill,revert}.ts; triage.ts
  imports the split modules (validateSnapshotPath + shared types/query in
  migrate-shared); migrate.ts deleted. Largest new file 416 l.
- doctor.ts (946 l) → doctor-{shared,github,local}.ts + 62-line thin main;
  all 13 inline require('node:fs') replaced with static ES imports.
- page.ts buildHtml: 15 positional params → single DashboardProps arg;
  ProjectMeta deduped into lib/types.ts (was duplicated in page.ts + dashboard.ts);
  both dashboard.ts call sites pass an object literal.
- cli/commands/dashboard.ts: 2 require('node:fs').unlinkSync → static unlinkSync import.

Zero behavior change: doctor --json byte-identical to baseline; typecheck, lint,
and 782 tests pass.
