+++ plugins/dev-core/skills/issue-triage/lib/migrate-audit.ts
@@ -1,0 +1,116 @@
+/**
+ * migrate-audit.ts — audit-schema subcommand for taxonomy migration.
+ */
+
+import {
+  GH_PROJECT_ID,
+  LANE_OPTIONS,
+  PRIORITY_OPTIONS,
+  SIZE_OPTIONS,
+  STATUS_OPTIONS,
+} from '../../shared/adapters/config-helpers'
+import { ghGraphQL } from '../../shared/adapters/github-adapter'
+
+// ---------------------------------------------------------------------------
+// Schema types
+// ---------------------------------------------------------------------------
+
+export interface FieldSchema {
+  id: string
+  name: string
+  options: Array<{ id: string; name: string }>
+}
+
+// ---------------------------------------------------------------------------
+// GraphQL query
+// ---------------------------------------------------------------------------
+
+export const PROJECT_FIELDS_QUERY = `
+  query($projectId: ID!) {
+    node(id: $projectId) {
+      ... on ProjectV2 {
+        fields(first: 50) {
+          nodes {
+            ... on ProjectV2SingleSelectField {
+              id
+              name
+              options { id name }
+            }
+          }
+        }
+      }
+    }
+  }
+`
+
+// ---------------------------------------------------------------------------
+// Expected field definitions
+// ---------------------------------------------------------------------------
+
+export const EXPECTED_FIELDS: Array<{ name: string; options: Record<string, string> }> = [
+  { name: 'Size', options: SIZE_OPTIONS },
+  { name: 'Lane', options: LANE_OPTIONS },
+  { name: 'Priority', options: PRIORITY_OPTIONS },
+  { name: 'Status', options: STATUS_OPTIONS },
+]
+
+// ---------------------------------------------------------------------------
+// auditSchema
+// ---------------------------------------------------------------------------
+
+export async function auditSchema(): Promise<void> {
+  // fix #10: wrap ghGraphQL in try/catch
+  let data: {
+    node: {
+      fields: {
+        nodes: Array<Partial<FieldSchema>>
+      }
+    }
+  }
+  try {
+    data = (await ghGraphQL(PROJECT_FIELDS_QUERY, { projectId: GH_PROJECT_ID })) as typeof data
+  } catch (err) {
+    const msg = err instanceof Error ? err.message : String(err)
+    console.error(`Error: audit-schema failed to query project. Check GH_PROJECT_ID and GH token. Underlying: ${msg}`)
+    process.exit(1)
+  }
+
+  const liveNodes = data.node.fields.nodes.filter((n): n is FieldSchema => Array.isArray((n as FieldSchema).options))
+
+  const liveByName = new Map<string, FieldSchema>(liveNodes.map((n) => [n.name, n]))
+
+  const diffs: string[] = []
+
+  for (const expected of EXPECTED_FIELDS) {
+    const liveField = liveByName.get(expected.name)
+
+    if (!liveField) {
+      diffs.push(`MISSING: ${expected.name}`)
+      continue
+    }
+
+    const liveOptionNames = new Set(liveField.options.map((o) => o.name))
+    const localOptionNames = new Set(Object.keys(expected.options))
+
+    for (const local of localOptionNames) {
+      if (!liveOptionNames.has(local)) {
+        diffs.push(`LOCAL_EXTRA: ${expected.name}.${local}`)
+      }
+    }
+
+    for (const live of liveOptionNames) {
+      if (!localOptionNames.has(live)) {
+        diffs.push(`LOCAL_MISSING: ${expected.name}.${live}`)
+      }
+    }
+  }
+
+  if (diffs.length > 0) {
+    for (const line of diffs) {
+      console.log(line)
+    }
+    process.exit(1)
+  }
+
+  console.log('audit-schema: OK — Size/Lane/Priority/Status match live project')
+}
